Fixed the venv usage inside a devcontainer
The devcontainer would complain that the `.venv` of my project was not
valid.
It turns out that python creates symlink from the bin folder to your
system's python (amongst other things).
The venv was created on my laptop. When used the symlink was pointing to
a path that just did not exist in my container.